requirements.txt
allure-pytest==2.9.45
selene==2.0.0rc4
pytest==7.4.4
selenium==4.38.0
webdriver-manager==3.8.6